pyproject.toml
[project]
name = "refineryai-document-intelligence-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.76.0",
    "numpy>=2.4.2",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "pytesseract>=0.3.13",
    "pytest>=9.0.2",
]
